backend/requirements.txt
fastapi>=0.115.0
sqlmodel>=0.0.22
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.4.0
